+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 ### Other Changes
 
 * Debug logging includes the address of the object that's writing a log entry.
+* Context expiration or cancellation when creating instances of `Session`, `Receiver`, and `Sender` no longer result in the potential for `Conn` to unexpectedly terminate.
+* Session channel and link handle exhaustion will now return `*ConnError` and `*SessionError` respectively, closing the respective `Conn` or `Session`.
 
 ## 0.19.1 (2023-03-31)